Fix printed bug in roll-independent results
The roll values in the display of observations with roll-independent
results were "wrong" as the printed roll was just a roll that was used for
the single roll-independent check, instead of a nominal roll corresponding
to the day in the table.  Printed results fixes for consistency, though this
should really have no impact.